move submit page layout to separate component
